### 2025-09-22 #9047 @editorialbot
[PRE REVIEW]: Graphorge: An open-source forge of graph neural networks
**Submitting author:** @BernardoFerreira (<a href="http://orcid.org/0000-0001-5956-3877">Bernardo P. Ferreira</a>)
**Repository:** https://github.com/bessagroup/graphorge
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
