### 2025-09-23 2a00974
Update installation instructions
