### 2025-09-23 edf5116
Added direct install option
