+++ setup.cfg
@@ -25,7 +25,7 @@ package_dir=
   =src
 packages=find:
 include_package_data = True
-python_requires = >=3.8
+python_requires = >=3.11
 install_requires = file: requirements.txt
 
 
