### 2025-09-23 31792c9
Updated scipy version
