+++ requirements.txt
@@ -4,7 +4,7 @@ numpy==1.26.4
 pandas==2.1.1
 pytest==7.1.2
 scikit_learn==1.3.2
-scipy==1.10.0
+scipy==1.11.4
 torch==2.4.1
 torch_geometric==2.6.1
 tqdm==4.65.0
