+++ environment_macos.yml
@@ -14,7 +14,7 @@ dependencies:
   - pandas=2.1.1
   - pytest=7.1.2
   - coverage=7.3.2
-  - scipy=1.10.0
+  - scipy=1.11.4
   - sphinx_rtd_theme=1.2.0
   - tqdm=4.65.0
   - shapely=2.0.2
