### 2025-09-23 b79c7c1
Removed unused shapely requirement
