+++ requirements.txt
@@ -8,7 +8,6 @@ scipy==1.11.4
 torch==2.4.1
 torch_geometric==2.6.1
 tqdm==4.65.0
-shapely==2.0.2
 torchinfo==1.8.0
 hydra-core==1.3.2
 hydra-nevergrad-sweeper==1.3.0.dev0
