+++ environment_macos.yml
@@ -17,7 +17,6 @@ dependencies:
   - scipy=1.11.4
   - sphinx_rtd_theme=1.2.0
   - tqdm=4.65.0
-  - shapely=2.0.2
   - hydra-core=1.3.2
   - torchinfo=1.8.0
   # Check:
