### 2025-09-24 750caf3
Removed pandas from core requirements
