+++ requirements.txt
@@ -1,7 +1,6 @@
 cycler==0.11.0
 matplotlib==3.8.0
 numpy==1.26.4
-pandas==2.1.1
 pytest==7.1.2
 scikit_learn==1.3.2
 scipy==1.11.4
