+++ environment.yml
@@ -11,7 +11,6 @@ dependencies:
   - cycler=0.11.0
   - matplotlib=3.8.0
   - numpy=1.26.4
-  - pandas=2.1.1
   - pytest=7.1.2
   - coverage=7.3.2
   - scipy=1.11.4
