+++ _sources/_collections/examples/mechanics/gnn_shell_buckling/gnn_shell_buckling.ipynb.txt
@@ -43,9 +43,10 @@
     "    %pip install torch-scatter -f https://data.pyg.org/whl/torch-2.4.1+cu124.html\n",
     "    %pip install torch-sparse -f https://data.pyg.org/whl/torch-2.4.1+cu124.html\n",
     "    %pip install torch_geometric==2.6.1 -f https://data.pyg.org/whl/torch-2.4.1+cu124.html\n",
-    "    \n",
+    "\n",
     "    # Install Graphorge\n",
-    "    %pip install git+https://github.com/bessagroup/graphorge.git\n",
+    "    %git clone https://github.com/bessagroup/graphorge.git\n",
+    "    %pip install -e ./graphorge\n",
     "except:\n",
     "    pass\n",
     "\n",
@@ -93,9 +94,16 @@
     "from graphorge.gnn_base_model.predict.prediction import predict\n",
     "\n",
     "# Make utilities available in the path\n",
-    "utils_path = (\n",
-    "    Path().resolve().parents[1] / \"utils\"\n",
-    ")\n",
+    "try:\n",
+    "    # Check if running in Google Colab\n",
+    "    import google.colab\n",
+    "    utils_path = (\n",
+    "      Path().resolve() / \"graphorge\" / \"benchmarks\" / \"utils\"\n",
+    "    )\n",
+    "except:\n",
+    "    utils_path = (\n",
+    "        Path().resolve().parents[1] / \"utils\"\n",
+    "    ) \n",
     "sys.path.insert(0, str(utils_path))\n",
     "\n",
     "# Import required utilities\n",
