+++ _sources/_collections/examples/cfd/gnn_porous_medium/gnn_porous_medium.ipynb.txt
@@ -39,7 +39,8 @@
     "    %pip install torch_geometric==2.6.1 -f https://data.pyg.org/whl/torch-2.4.1+cu124.html\n",
     "    \n",
     "    # Install Graphorge\n",
-    "    %pip install git+https://github.com/bessagroup/graphorge.git\n",
+    "    %git clone https://github.com/bessagroup/graphorge.git\n",
+    "    %pip install -e ./graphorge\n",
     "except:\n",
     "    pass"
    ]
@@ -54,7 +55,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 2,
+   "execution_count": null,
    "id": "2af543bc",
    "metadata": {},
    "outputs": [],
@@ -79,9 +80,16 @@
     "from graphorge.gnn_base_model.predict.prediction import predict\n",
     "\n",
     "# Make utilities available in the path\n",
-    "utils_path = (\n",
-    "    Path().resolve().parents[1] / \"utils\"\n",
-    ")\n",
+    "try:\n",
+    "    # Check if running in Google Colab\n",
+    "    import google.colab\n",
+    "    utils_path = (\n",
+    "      Path().resolve() / \"graphorge\" / \"benchmarks\" / \"utils\"\n",
+    "    )\n",
+    "except:\n",
+    "    utils_path = (\n",
+    "        Path().resolve().parents[1] / \"utils\"\n",
+    "    ) \n",
     "sys.path.insert(0, str(utils_path))\n",
     "\n",
     "# Import required utilities\n",
