### 2025-09-25 e2cf171
Updated shell benchmark
