+++ benchmarks/utils/utils.py
@@ -9,6 +9,7 @@ import pickle
 
 from itertools import count
 
+
 class DownloadProgressBar(tqdm):
     """
     Provides a progress bar compatible with the urlretrieve hook.
@@ -67,7 +68,9 @@ def download_file(file, base_url, dest_path, overwrite=False):
     # Check if the file already exists and use it if overwrite is False
     if (dest_path / file).exists() and not overwrite:
         # Logging
-        logging.info(msg=f"Using '{file}' cached in in {dest_path}")
+        logging.info(
+            msg=f"Using '{file}' cached in in {dest_path.relative_to(Path.cwd())}"
+        )
         return
 
     # Create the destination directory if it doesn't exist
@@ -95,6 +98,7 @@ def download_file(file, base_url, dest_path, overwrite=False):
         # total_size's units does not seem to be in bytes
         pbar.total = pbar.n
 
+
 def get_critical_buckling_wavelength(poisson, shell_radius, shell_thickness):
     """Compute shell critical buckling wavelength.
 
@@ -119,7 +123,6 @@ def get_critical_buckling_wavelength(poisson, shell_radius, shell_thickness):
         * ((12 * (1 - poisson**2)) ** (-1 / 4))
         * ((shell_radius * shell_thickness) ** (1 / 2))
     )
-    # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     return critical_bw
 
 
@@ -179,8 +182,6 @@ def parse_tensorflow_dataset(dataset_name, dataset_directory):
             pickle.dump(sample_data, file)
 
 
-
-
 def _parse_tensorflow_sample(sample, metadata):
     """
     Parses a single sample from the TensorFlow dataset.
@@ -207,15 +208,11 @@ def _parse_tensorflow_sample(sample, metadata):
         )
         raise ImportError(err_msg)
 
-    feature_dict = {
-        k: tf.io.VarLenFeature(tf.string) for k in metadata["field_names"]
-    }
+    feature_dict = {k: tf.io.VarLenFeature(tf.string) for k in metadata["field_names"]}
     features = tf.io.parse_single_example(sample, feature_dict)
     sample_data = {}
     for key, field in metadata["features"].items():
-        data = tf.io.decode_raw(
-            features[key].values, getattr(tf, field["dtype"])
-        )
+        data = tf.io.decode_raw(features[key].values, getattr(tf, field["dtype"]))
         data = tf.reshape(data, field["shape"])
         if field["type"] == "static":
             data = data[0]
@@ -225,6 +222,115 @@ def _parse_tensorflow_sample(sample, metadata):
         sample_data[key] = data.numpy()
 
     return sample_data
+
+
+plotly_layout = dict(
+    template="plotly_white",
+    font_family="Arial",
+    font_size=10,
+    xaxis_showgrid=False,
+    xaxis_ticks="inside",
+    xaxis_mirror="allticks",
+    xaxis_zeroline=False,
+    xaxis_showline=True,
+    xaxis_linecolor="black",
+    yaxis_showgrid=False,
+    yaxis_ticks="inside",
+    yaxis_mirror="allticks",
+    yaxis_zeroline=False,
+    yaxis_showline=True,
+    yaxis_linecolor="black",
+    colorway=["#4477aa", "#ef6f7f", "#288b39", "#288b39", "#aa3377", "#000000"],
+    xaxis_title_standoff=3,
+    yaxis_title_standoff=3,
+    margin=dict(l=40, r=10, t=10, b=40),
+    yaxis_automargin=True,
+)
+"""A Plotly layout dictionary for consistent figure styling."""
+
+
[…]
