### 2025-10-02 9ea6d63
MGN benchmark
