+++ benchmarks/utils/utils.py
@@ -9,6 +9,7 @@ import pickle
 
 from itertools import count
 from graphorge.gnn_base_model.data.graph_data import GraphData
+from torch_geometric.data import Data
 
 
 class DownloadProgressBar(tqdm):
@@ -250,13 +251,13 @@ plotly_layout = dict(
 """A Plotly layout dictionary for consistent figure styling."""
 
 
-def plot_interactive_graph(graph_data: GraphData, node_size: int = 10):
+def plot_interactive_graph(graph_data: GraphData | Data, node_size: int = 10):
     """
     Plots the graph using Plotly.
 
     Parameters
     ----------
-    graph_data : GraphData
+    graph_data : GraphData | Data
         The graph data to plot.
     """
     import plotly.graph_objects as go
@@ -265,7 +266,10 @@ def plot_interactive_graph(graph_data: GraphData, node_size: int = 10):
     figure = go.Figure()
 
     # Extract node coordinates
-    node_coords = graph_data.get_nodes_coords()
+    if isinstance(graph_data, GraphData):
+        node_coords = graph_data.get_nodes_coords()
+    else:
+        node_coords = graph_data.pos
 
     if node_coords.shape[1] == 3:
         # 3D graph
@@ -273,8 +277,13 @@ def plot_interactive_graph(graph_data: GraphData, node_size: int = 10):
     else:
         is_3d = False
 
+    # Extract edge indexes
+    if isinstance(graph_data, GraphData):
+        edge_indexes = graph_data.get_graph_edges_indexes()
+    else:
+        edge_indexes = graph_data.edge_index.T
     # Prepare edge coordinates by indexing node coordinates with edge indexes
-    edge_x = node_coords[graph_data.get_graph_edges_indexes()][:, :, 0].ravel()
+    edge_x = node_coords[edge_indexes][:, :, 0].ravel()
 
     # We need to insert None values between edges to avoid connecting them
     # in the plotly scatter3d while adding only one trace
@@ -282,22 +291,32 @@ def plot_interactive_graph(graph_data: GraphData, node_size: int = 10):
     edge_x = np.insert(edge_x, insert_index, None)
 
     # Prepare y edge coordinates
-    edge_y = node_coords[graph_data.get_graph_edges_indexes()][:, :, 1].ravel()
+    edge_y = node_coords[edge_indexes][:, :, 1].ravel()
     edge_y = np.insert(edge_y, insert_index, None)
 
     # Prepare z edge coordinates
     if is_3d:
-        edge_z = node_coords[graph_data.get_graph_edges_indexes()][:, :, 2].ravel()
+        edge_z = node_coords[edge_indexes][:, :, 2].ravel()
         edge_z = np.insert(edge_z, insert_index, None)
 
     # Prepare hovertemplate for edge features, if any
-    edge_feature_names = graph_data.get_metadata().get("edge_feature_names", [])
+    if isinstance(graph_data, GraphData):
+        edge_feature_names = graph_data.get_metadata().get("edge_feature_names", [])
+    else:
+        edge_feature_names = graph_data.metadata.get("edge_feature_names", [])
+
     hovertemplate_edge_features = "".join(
         [
             f"<br>{name}: %{{customdata[{i}]:.2f}}"
             for i, name in enumerate(edge_feature_names)
         ]
     )
+
+    # Extract edge features, if any
+    if isinstance(graph_data, GraphData):
+        edge_features = graph_data.get_edge_features_matrix()
+    else:
+        edge_features = graph_data.edge_attr
 
     # Add edges
     if is_3d:
@@ -314,15 +333,9 @@ def plot_interactive_graph(graph_data: GraphData, node_size: int = 10):
 
         if hovertemplate_edge_features:
             # Add invisible scatter for edge features hover at mid edge position
[…]
