### 2025-10-08 0a47f15
Release MGN benchmark (#35)
* Updated shell benchmark

* Updated porous benchmark

* Implement cylinder flow dataset

* More explicit doctring

* Propagate device type

* MGN benchmark

* Cleaning

* Update color map

* Updated error colormap
