+++ _sources/_collections/examples/readme.md.txt
@@ -17,4 +17,5 @@ The provided examples can be ran:
 |Example|Code|Field|Concepts covered|
 | --- | --- | --- | --- |
 |[Shell buckling](./mechanics/gnn_shell_buckling/gnn_shell_buckling)|[Github](https://github.com/bessagroup/graphorge/tree/main/benchmarks/mechanics/gnn_shell_buckling/gnn_shell_buckling.ipynb)|Mechanics|Graphorge basics, regression on a global target.|
-|[Permeability of a porous medium](./cfd/gnn_porous_medium/gnn_porous_medium)|[Github](https://github.com/bessagroup/graphorge/tree/main/benchmarks/cfd/gnn_porous_medium/gnn_porous_medium.ipynb)|CFD|Graphorge basics, regression on a global target.|
+|[Flow rate in a porous medium](./cfd/gnn_porous_medium/gnn_porous_medium)|[Github](https://github.com/bessagroup/graphorge/tree/main/benchmarks/cfd/gnn_porous_medium/gnn_porous_medium.ipynb)|CFD|Graphorge basics, regression on a global target.|
+|[MeshGraphNet for turbulent flow around a cylinder](./cfd/mgn_cylinder_flow/mgn_cylinder_flow)|[Github](https://github.com/bessagroup/graphorge/tree/main/benchmarks/cfd/mgn_cylinder_flow/mgn_cylinder_flow.ipynb)|CFD|Graphorge basics, regression on node targets.|
