### 2025-11-17 #9047 @asogaard
> 👋 [@asogaard](https://github.com/asogaard) & [@diningphil](https://github.com/diningphil), would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @jromanowska,
I am unfortunately not able to review this submission. Is there somewhere I can indicate this for further submissions, to save you the hassle of reaching out?
