### 2026-01-11 #9453 @MSamane
Hi @BernardoFerreira
I noticed that the link provided in the benchmarks section does not connect to an existing page: https://github.com/bessagroup/graphorge/tree/main/benchmarks
