### 2026-01-13 #9453 @MSamane
@GuillaumeBroggi
To improve usability, I suggest adding a brief “Getting started” or “Usage overview” section in the paper. This section could summarize the typical workflow and point readers to the relevant examples or entry points in the GitHub repository.
