### 2026-01-14 #9453 @jromanowska
> [@GuillaumeBroggi](https://github.com/GuillaumeBroggi) To improve usability, I suggest adding a brief “Getting started” or “Usage overview” section in the paper. This section could summarize the typical workflow and point readers to the relevant examples or entry points in the GitHub repository.

Thank you for the work so far, @MSamane and @GuillaumeBroggi. I need to chip in here. JOSS' paper is meant to be short and we do not want any additional sections. The section you suggested, @MSamane, would fit best in README or Documentation for the software.
