+++ _sources/_collections/examples/mechanics/gnn_shell_buckling/gnn_shell_buckling.ipynb.txt
@@ -49,7 +49,7 @@
                 "    %pip install torch_geometric==2.6.1 -f https://data.pyg.org/whl/torch-2.4.1+cu124.html\n",
                 "\n",
                 "    # Install Graphorge\n",
-                "    %git clone https://github.com/bessagroup/graphorge.git\n",
+                "    !git clone https://github.com/bessagroup/graphorge.git\n",
                 "    %pip install -e ./graphorge\n",
                 "except ImportError:\n",
                 "    pass\n",
@@ -81,8 +81,6 @@
             "metadata": {},
             "outputs": [],
             "source": [
-                "%load_ext autoreload\n",
-                "%autoreload 2\n",
                 "# Standard\n",
                 "import logging\n",
                 "from pathlib import Path\n",
@@ -111,8 +109,7 @@
                 "try:\n",
                 "    # Check if running in Google Colab\n",
                 "    import google.colab\n",
-                "\n",
-                "    utils_path = Path().resolve() / \"graphorge\" / \"benchmarks\" / \"utils\"\n",
+                "    utils_path = Path(\"graphorge/benchmarks/utils\")\n",
                 "except:\n",
                 "    utils_path = Path().resolve().parents[1] / \"utils\"\n",
                 "sys.path.insert(0, str(utils_path))\n",
@@ -6659,7 +6656,7 @@
             "name": "python",
             "nbconvert_exporter": "python",
             "pygments_lexer": "ipython3",
-            "version": "3.12.3"
+            "version": "3.11.9"
         },
         "title": "A GNN surrogate to estimate the knock-down factor of imperfect shells",
         "widgets": {
