+++ _sources/_collections/examples/cfd/mgn_cylinder_flow/mgn_cylinder_flow.ipynb.txt
@@ -52,7 +52,7 @@
                 "    %pip install torch_geometric==2.6.1 -f https://data.pyg.org/whl/torch-2.4.1+cu124.html\n",
                 "    \n",
                 "    # Install Graphorge\n",
-                "    %git clone https://github.com/bessagroup/graphorge.git\n",
+                "    !git clone https://github.com/bessagroup/graphorge.git\n",
                 "    %pip install -e ./graphorge\n",
                 "except:\n",
                 "    pass\n",
@@ -94,8 +94,6 @@
             "metadata": {},
             "outputs": [],
             "source": [
-                "%load_ext autoreload\n",
-                "%autoreload 2\n",
                 "# Standard\n",
                 "import json\n",
                 "import logging\n",
@@ -127,9 +125,7 @@
                 "try:\n",
                 "    # Check if running in Google Colab\n",
                 "    import google.colab\n",
-                "    utils_path = (\n",
-                "      Path().resolve() / \"graphorge\" / \"benchmarks\" / \"utils\"\n",
-                "    )\n",
+                "    utils_path = Path(\"graphorge/benchmarks/utils\")\n",
                 "except:\n",
                 "    utils_path = (\n",
                 "        Path().resolve().parents[1] / \"utils\"\n",
@@ -6659,7 +6655,6 @@
                 "# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n",
                 "# Load training datasets\n",
                 "train_dataset = torch.load(directories[\"training\"] / \"cylinder_flow_dataset.pt\")\n",
-                "train_dataset.to(device_type)\n",
                 "validation_dataset = torch.load(directories[\"validation\"] / \"cylinder_flow_dataset.pt\")\n",
                 "\n",
                 "training_parameters = dict(\n",
