### 2026-01-16 #9453 @MSamane
Thank you @GuillaumeBroggi, now, I was able to open and run it successfully. For the other examples, are you planning to make the dataset for the first two examples publicly available later?

Thanks also for the clarification about the documents. I've already reviewed them and they look very useful.
