### 2026-01-19 #9453 @marjanalbooyeh
Hi @GuillaumeBroggi , minor issue in benchmarks readme file.

https://github.com/bessagroup/graphorge/issues/36
