### 2026-01-19 #9453 @marjanalbooyeh
Hi @GuillaumeBroggi  and @jromanowska ,

Apologies for the delay.  Quick notes from my review:

- Installation instructions are complete and easy to follow. I was able to follow the documented installation steps successfully.
- I was able to run the third example Jupyter notebook (the one with the public dataset) end-to-end and reproduce the reported results. The example notebooks are clear, and the code is well commented.
- The documentation is clear and provides sufficient guidance for getting started, understanding the package structure, and running the provided examples.
- The paper provides a clear summary and statement of need, and the repository is organized in a way that makes it easy for new users to understand the workflow and architecture.

### Minor Suggestion (not an acceptance blocker)

Even though the authors mention the package can be used across many domains, the current examples are all from computational mechanics/CFD. It would be nice to include **at least one simple example from another are**a (e.g., a well-known graph benchmark like molecule property prediction or citation-network node classification) to better support that claim and help attract a wider audience.
