### 2026-01-27 #9453 @GuillaumeBroggi
Dear @marjanalbooyeh and @jromanowska,

Thank you for the effort you put into reviewing our submission, we greatly appreciate it.

> ### Minor Suggestion (not an acceptance blocker)
> Even though the authors mention the package can be used across many domains, the current examples are all from computational mechanics/CFD. It would be nice to include **at least one simple example from another are**a (e.g., a well-known graph benchmark like molecule property prediction or citation-network node classification) to better support that claim and help attract a wider audience.

We claim in the paper's summary that

> Although originally developed for computational mechanics, Graphorge's core functionality is general and can be applied across a wide range of domains

because Graphorge is a tool to implement and experiment with graph neural networks (GNNs). Graphs are agnostic to the data they encode, and GNNs have been used in a wide range of applications, from mechanics to molecular properties predictions.

The examples presented with the package focus on computational mechanics, reflecting our expertise and research interests. We could nuance the claim as follows:

> Although originally developed for computational mechanics, Graphorge's core functionality is built around graph neural networks, which have been successfully applied across a wide range of scientific and engineering domains.

Would that be more acceptable?
