### 2026-02-09 #9453 @jromanowska
@BernardoFerreira, @GuillaumeBroggi - I will now check the repo and paper before I can recommend an acceptance. Could you go through the points above, under "Additional Author Tasks" and report back?
