### 2026-02-11 #9453 @jromanowska
@BernardoFerreira, @GuillaumeBroggi, I've read the paper and I have no comments apart from one grammatical mistake: in line 27, you wrote "for researchers and studnets _that_" - this should probably be _who_.

Just one question to the author list: who is the last author and how did they contribute to the software?
