### 2026-02-11 80d5667
Implement JOSS review comments
