+++ docs/joss_paper/paper.md
@@ -29,7 +29,7 @@ bibliography: references.bib
 
 # Summary
 
-[Graphorge](https://github.com/bessagroup/graphorge) is an open-source Python package built on [PyTorch](https://pytorch.org/) that streamlines the development and evaluation of graph neural networks. It provides a complete workflow encompassing data pre-processing, dataset management, model training, prediction, and result post-processing. Graphorge includes a fully implemented, highly customizable example architecture to demonstrate practical use, and its code is thoroughly documented and commented, making it especially accessible to researchers new to implementing graph neural networks. Although originally developed for computational mechanics, Graphorge’s core functionality is general and can be applied across a wide range of domains.
+[Graphorge](https://github.com/bessagroup/graphorge) is an open-source Python package built on [PyTorch](https://pytorch.org/) that streamlines the development and evaluation of graph neural networks. It provides a complete workflow encompassing data pre-processing, dataset management, model training, prediction, and result post-processing. Graphorge includes a fully implemented, highly customizable example architecture to demonstrate practical use, and its code is thoroughly documented and commented, making it especially accessible to researchers new to implementing graph neural networks. Although originally developed for computational mechanics, Graphorge's core functionality is built around graph neural networks, which have been successfully applied across a wide range of scientific and engineering domains.
 
 ![Logo of [Graphorge](https://github.com/bessagroup/graphorge). \label{fig:graphorge_logo_horizontal_white}](graphorge_logo_horizontal_white.png)
 
@@ -37,7 +37,7 @@ bibliography: references.bib
 
 Graph neural networks have emerged as a powerful modeling tool for data with relational or geometric structure. Existing graph neural network libraries tend to fall into two categories. On the one hand, many research-specific implementations are minimally documented and tightly tailored to particular benchmarks. While this may enable results reproducibility, code comprehension is often limited, as well as its customization and generalization to different applications. On the other hand, general purpose frameworks such as [PyTorch Geometric](https://pytorch-geometric.readthedocs.io/en/latest/) [@fey:2019a] and [Deep Graph Library](https://www.dgl.ai/) [@wang:2020a] offer robust, high-performance platforms to implement graph neural networks, usually providing multiple backend support and integration of state-of-the-art scientific contributions.
 
-Graphorge is not intended as an alternative to general-purpose frameworks. Instead, it is designed as a practical and educational tool for researchers and students that aim to understand and be able to implement a full Graph Neural Network pipeline. While only including a single, highly customizable architecture [@battaglia:2018a] to demonstrate practical use, every single module in Graphorge is extensively documented and commented, aiming to maximize code comprehension. Moreover, rather than abstracting the workflow behind opaque interfaces, Graphorge provides deliberatly fully functional, modular scripts for each stage -- from pre-processing and dataset handling to model training and post-processing -- making it an ideal extensible, starting point for those looking to implement or customize graph neural networks in a research environment.
+Graphorge is not intended as an alternative to general-purpose frameworks. Instead, it is designed as a practical and educational tool for researchers and students who aim to understand and be able to implement a full Graph Neural Network pipeline. While only including a single, highly customizable architecture [@battaglia:2018a] to demonstrate practical use, every single module in Graphorge is extensively documented and commented, aiming to maximize code comprehension. Moreover, rather than abstracting the workflow behind opaque interfaces, Graphorge provides deliberately fully functional, modular scripts for each stage -- from pre-processing and dataset handling to model training and post-processing -- making it an ideal extensible, starting point for those looking to implement or customize graph neural networks in a research environment.
 
 ![Example of Graph Neural Network model with an encoder-processor-decoder architecture.](graphorge_overview.png)
 
