### 2026-02-11 #9453 @GuillaumeBroggi
@jromanowska I am on it. I have already pushed bessagroup/graphorge@80d5667b591131355bb8453db4ef2edb0eef98bc to implement the edit approved by @marjanalbooyeh, the typo you spotted (and another one I spotted).

The last author is @mabessa. He is the PI behind the research that led to this piece of software.
