### 2026-02-14 #9453 @GuillaumeBroggi
@jromanowska, I have made a new release with the latest changes from the review: [1.0.1](https://github.com/bessagroup/graphorge/releases/tag/1.0.1)

The repository is archived in Zenodo: [10.5281/zenodo.18643444](https://doi.org/10.5281/zenodo.18643444)
